Add AWS Lambda steps to Github Action
After the merge to master, we want to deploy the function to AWS Lambda.
Since we're working on their build machines, we need to first install
the dotnet lambda tool, then publish the build using dotnet lambda.

For security, we pull the various configuration options from Github
secrets so they're not in source control.  To support this, we need to
add the following secrets:

 * AWS_ACCESS_KEY_ID
 * AWS_SECRET_KEY
 * AWS_REGION

Once these are set, this will publish changes to the lambda function on
merge to master.